pyproject.toml
[project]
name = "scriptorium"
version = "0.1.0"
description = "Personal book tracking and review application"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  {name = "Tobias Kunze", email = "r@rixx.de"},
]

dependencies = [
  "beautifulsoup4~=4.14",
  "Django~=6.0",
  "django-context-decorator~=1.6",
  "django-formtools~=2.5",
  "gunicorn",
  "inquirer~=3.4",
  "Jinja2~=3.1",
  "Markdown~=3.10",
  "MarkupSafe~=3.0",
  "networkx~=3.6",
  "numpy~=2.4",
  "Pillow~=12.2",
  "pygal~=3.1",
  "python-dateutil~=2.9",
  "python-frontmatter~=1.1",
  "requests~=2.33",
  "scikit-learn~=1.8",
  "smartypants~=2.0",
  "tqdm~=4.67",
  "Unidecode~=1.4",
  "whitenoise~=6.12",
]

[project.optional-dependencies]
dev = [
  "ruff",
]

[project.urls]
Homepage = "https://books.rixx.de"
Repository = "https://github.com/rixx/scriptorium"

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "wheel"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["scriptorium*"]

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN",     # We do not do annotations
  "ARG",     # Unused method arguments are often ok
  "C901",    # We do do complexity
  "COM812",  # Trailing commas are formatter territory
  "D",       # We do not enforce docstyles
  "DJ001",   # We live in a state of sin
  "E501",    # Long lines are formatter territory
  "EM",      # Formatting strings is fine, actually
  "FBT",     # Booleans are fine, actually
  "FIX",     # TODO comments are fine, actually
  "PLR",     # Refactoring: Too many branches, arguments, statements, etc
  "RET502",  # Just returning is fine, actually
  "RET503",  # Just not returning is fine, actually
  "RUF012",  # Mutable default in class attribute is just how django rolls
  "RUF100",  # Unused noqa; problematic when run in isolation
  "TD",      # TODO formatting (author, issue link, colon) is not useful here
  "TRY003",  # Error messages are fine, actually
]
allowed-confusables = ["–", "'", "×"]

[tool.ruff.lint.per-file-ignores]
"src/manage.py" = [
  "INP001",  # Standard Django manage.py is not in a package
  "PLC0415", # Standard Django manage.py pattern
]
"src/helpers.py" = [
  "F403",    # Star imports are fine in a helper script
  "F405",    # Ditto
  "INP001",  # Not in a package
  "T201",    # print() is the primary output mechanism
]
"src/scriptorium/main/management/**.py" = [
  "T201",    # print() is fine in management commands
]
"src/tests/**.py" = [
  "INP001",  # Tests are not in a package
  "S101",    # Assertions are the whole point
  "S106",    # Hardcoded passwords are fine in tests
  "PT011",   # Broad pytest.raises is fine
  "SLF001",  # Accessing private members is fine in tests
]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.flake8-self]
extend-ignore-names = ["_base_manager", "_default_manager", "_meta"]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-django>=4.12.0",
    "pytest-xdist>=3.6.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "scriptorium.settings"
pythonpath = ["src"]
python_files = ["test_*.py"]
testpaths = ["src/tests"]
filterwarnings = [
  "ignore::DeprecationWarning",
]

[tool.coverage.run]
branch = true
omit = [
  "*/__init__.py",
  "*/__main__.py",
  "*/helpers.py",
  "*/manage.py",
  "*/management/commands/*",
  "*/migrations/*",
  "*/settings*",
  "*/tests/*",
  "*/urls.py",
  "*/wsgi.py",
  "*/asgi.py",
]
relative_files = true

[tool.coverage.report]
fail_under = 100
exclude_also = [
  "def __repr__",
  "def __str__",
  "if settings.DEBUG",
  "NOQA",
  "noqa",
  "NotImplementedError",
]
